requirements.txt
python-telegram-bot>=21.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
PyPDF2>=4.0.0
python-docx>=0.8.11
openpyxl>=3.11.0
python-dotenv>=1.0.0
